Change confidence score to float (e.g. 4.2/5)

+++ backend/app/db/models/draft.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, Text, Integer, ForeignKey, DateTime, func
+from sqlalchemy import String, Text, Float, ForeignKey, DateTime, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base_class import Base
 
@@ -14,7 +14,7 @@ class Draft(Base):
     approved_at:    Mapped[DateTime | None] = mapped_column(DateTime(timezone=True))
     sent_at:        Mapped[DateTime | None] = mapped_column(DateTime(timezone=True))
     gmail_draft_id:   Mapped[str | None]     = mapped_column(String(255))
-    confidence_score: Mapped[int | None]     = mapped_column(Integer)
+    confidence_score: Mapped[float | None]   = mapped_column(Float)
     created_at:     Mapped[DateTime]       = mapped_column(DateTime(timezone=True), server_default=func.now())
 
     email: Mapped['Email'] = relationship('Email', back_populates='draft')  # noqa: F821
